split Upgrade Box from TroopButton
